move xspress3 tests requiring hardware to a separate file
